# frontend/src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Glory RS API",
    "description": "Template API — Rust + Axum + OpenAPI",
    "contact": {
      "name": "1ndoryu"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "Sesión cerrada"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Usuario autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Usuario registrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "operationId": "profile",
        "responses": {
          "200": {
            "description": "Perfil del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      },
      "put": {
        "tags": [
          "profile"
        ],
        "operationId": "update_profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Perfil actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "403": {
            "description": "CSRF inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
    "/api/teams": {
      "get": {
        "tags": [
          "teams"
        ],
        "operationId": "get_team",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Equipo del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamOverview"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/teams/pending-count": {
      "get": {
        "tags": [
          "teams"
        ],
        "operationId": "pending_count",
        "responses": {
          "200": {
            "description": "Solicitudes pendientes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingTeamCount"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/teams/requests": {
      "post": {
        "tags": [
          "teams"
        ],
        "operationId": "send_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Solicitud creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamConnection"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/teams/requests/{id}": {
      "put": {
        "tags": [
          "teams"
        ],
        "operationId": "respond_request",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la solicitud",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamResponseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Solicitud respondida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamConnection"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/teams/{id}": {
      "delete": {
        "tags": [
          "teams"
        ],
        "operationId": "remove_connection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la conexión",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Conexión eliminada"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "type": "object",
        "description": "Response pública después de autenticarse. La sesión viaja únicamente en cookies.",
        "required": [
          "user"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Estructura de respuesta de error expuesta en la API",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Tipo de error (`not_found`, `unauthorized`, etc.)"
          },
          "message": {
            "type": "string",
            "description": "Mensaje legible para el usuario"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "description": "Request body para iniciar sesión",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "PendingTeamCount": {
        "type": "object",
        "required": [
          "pending"
        ],
        "properties": {
          "pending": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "description": "Request body para registrar un nuevo usuario",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "TeamConnection": {
        "type": "object",
        "required": [
          "id",
          "status",
          "requestedAt",
          "email",
          "isMine"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isMine": {
            "type": "boolean"
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time"
          },
          "respondedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "user": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TeamUser"
              }
            ],
            "nullable": true
          }
        }
      },
      "TeamCounts": {
        "type": "object",
        "required": [
          "received",
          "sent",
          "members"
        ],
        "properties": {
          "members": {
            "type": "integer",
            "format": "int64"
          },
          "received": {
            "type": "integer",
            "format": "int64"
          },
          "sent": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "TeamMember": {
        "type": "object",
        "required": [
          "id",
          "connectionId",
          "user"
        ],
        "properties": {
          "connectedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "connectionId": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user": {
            "$ref": "#/components/schemas/TeamUser"
          }
        }
      },
      "TeamOverview": {
        "type": "object",
        "required": [
          "received",
          "sent",
          "members",
          "counts",
          "page",
          "perPage",
          "hasMore"
        ],
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/TeamCounts"
          },
          "hasMore": {
            "type": "boolean"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamMember"
            }
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "perPage": {
            "type": "integer",
            "format": "int64"
          },
          "received": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamConnection"
            }
          },
          "sent": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamConnection"
            }
          }
        }
      },
      "TeamRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "TeamResponseRequest": {
        "type": "object",
        "required": [
          "action"
        ],
        "properties": {
          "action": {
            "type": "string"
          }
        }
      },
      "TeamUser": {
        "type": "object",
        "required": [
          "id",
          "displayName",
          "email"
        ],
        "properties": {
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "UpdateProfileRequest": {
        "type": "object",
        "description": "Cambios permitidos sobre el perfil público del usuario.",
        "required": [
          "display_name"
        ],
        "properties": {
          "avatar_url": {
            "type": "string",
            "nullable": true
          },
          "display_name": {
            "type": "string"
          }
        }
      },
...
      "UserResponse": {
        "type": "object",
        "description": "Response público de usuario — sin datos sensibles",
        "required": [
          "id",
          "email",
          "display_name",
          "created_at"
        ],
        "properties": {
          "avatar_url": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    },
    "securitySchemes": {
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id"
      }
    }
  }
}